refactor(www): dissolve components/marketing, drop dead code
components/marketing/ mixed two genuinely-marketing leaves (announcement,
rotating-text) with showcase/, which is a shared product surface consumed
by the home page, the /create preview, and the "Open in v0" bundle. Move
showcase/ to components/showcase/ and the two leaves to components/ root;
remove the now-empty marketing/ folder.

Also remove two dead hooks (use-debounce, use-google-font-loader; zero
importers) and rename config/site.tsx to .ts (pure data, no JSX).

Update all importers, the showcase-bundle build scripts, and the
Open-in-v0 page template; regenerate showcase-bundle.ts.
