+++ www/src/components/showcase/cards-grid.tsx
@@ -1,21 +1,21 @@
 import { cn } from '@/registry/lib/utils'
-import { AccountMenu } from '@/components/marketing/showcase/account-menu'
-import { AskAi } from '@/components/marketing/showcase/ask-ai'
-import { Booking } from '@/components/marketing/showcase/booking'
-import { ColorEditorCard } from '@/components/marketing/showcase/color-editor'
-import { ComputerUse } from '@/components/marketing/showcase/computer-use'
-import { CookiePreferences } from '@/components/marketing/showcase/cookie-preferences'
-import { Faq } from '@/components/marketing/showcase/faq'
-import { Filters } from '@/components/marketing/showcase/filters'
-import { InviteMembers } from '@/components/marketing/showcase/invite-members'
-import { LoginForm } from '@/components/marketing/showcase/login-form'
-import { Notifications } from '@/components/marketing/showcase/notifications'
-import { Payment } from '@/components/marketing/showcase/payment'
-import { Shortcuts } from '@/components/marketing/showcase/shortcuts'
-import { Storage } from '@/components/marketing/showcase/storage'
-import { TeamName } from '@/components/marketing/showcase/team-name'
-import { TwoFactor } from '@/components/marketing/showcase/two-factor'
-import { UploadAvatar } from '@/components/marketing/showcase/upload-avatar'
+import { AccountMenu } from '@/components/showcase/account-menu'
+import { AskAi } from '@/components/showcase/ask-ai'
+import { Booking } from '@/components/showcase/booking'
+import { ColorEditorCard } from '@/components/showcase/color-editor'
+import { ComputerUse } from '@/components/showcase/computer-use'
+import { CookiePreferences } from '@/components/showcase/cookie-preferences'
+import { Faq } from '@/components/showcase/faq'
+import { Filters } from '@/components/showcase/filters'
+import { InviteMembers } from '@/components/showcase/invite-members'
+import { LoginForm } from '@/components/showcase/login-form'
+import { Notifications } from '@/components/showcase/notifications'
+import { Payment } from '@/components/showcase/payment'
+import { Shortcuts } from '@/components/showcase/shortcuts'
+import { Storage } from '@/components/showcase/storage'
+import { TeamName } from '@/components/showcase/team-name'
+import { TwoFactor } from '@/components/showcase/two-factor'
+import { UploadAvatar } from '@/components/showcase/upload-avatar'
 
 // One masonry cell. The vertical gap between cards is the cell's *padding*-bottom,
 // never a margin: in a CSS multi-column layout WebKit/Safari leaks a child's
