+++ www/src/components/showcase/cards.tsx
@@ -5,9 +5,9 @@ import { useMemo, useState } from 'react'
 import { useMounted } from '@/hooks/use-mounted'
 import { DEFAULT_COLOR_CONFIG } from '@/registry/theme'
 import type { Density } from '@/registry/types'
-import { CardsGrid } from '@/components/marketing/showcase/cards-grid'
-import { ShowcaseCustomizer } from '@/components/marketing/showcase/showcase-customizer'
-import { SkeletonRail } from '@/components/marketing/showcase/skeleton-cards'
+import { CardsGrid } from '@/components/showcase/cards-grid'
+import { ShowcaseCustomizer } from '@/components/showcase/showcase-customizer'
+import { SkeletonRail } from '@/components/showcase/skeleton-cards'
 import { DesignSystemProvider } from '@/modules/core/styles'
 import { encodePreset } from '@/modules/create/preset'
