+++ www/src/components/showcase/computer-use.tsx
