+++ www/src/components/showcase/shortcuts.tsx
