--- www/src/hooks/use-debounce.ts
@@ -1,22 +1,0 @@
-import React from 'react'
-
-export function useDebounce<T>(value: T, delay: number): T {
-  const [debouncedValue, setDebouncedValue] = React.useState(value)
-
-  React.useEffect(() => {
-    if (delay <= 0) {
-      setDebouncedValue(value)
-      return
-    }
-
-    const handler = setTimeout(() => {
-      setDebouncedValue(value)
-    }, delay)
-
-    return () => {
-      clearTimeout(handler)
-    }
-  }, [delay, value])
-
-  return debouncedValue
-}
